* 'master' of github.com:z-song/laravel-admin: (111 commits)
  convert the route to kebab-case
  Update Form.php
  Fix #3496
  wip
  修正hasMany内->table新增事件污染到hasMany新增
  Fix issue for set default value of daterange form field
  Create FUNDING.yml
  wip
  Apply fixes from StyleCI
  Fix refresh btn issue
  Form/Listbox add height method
  Apply fixes from StyleCI
  Apply fixes from StyleCI
  Modify version number
  Multiple step form
  Change hasCheckAll to canCheckAll
  Add Field\Checkbox add `hasCheckAll()` method
  Apply fixes from StyleCI
  wip
  Update Filter.php
  ...

# Conflicts:
#	src/Form/Field.php
#	src/Form/Field/DateRange.php
#	src/Form/Field/UploadField.php
#	src/Grid.php

+++ src/Form/Field/UploadField.php
@@ -45,6 +45,13 @@ trait UploadField
     protected $useSequenceName = false;
 
     /**
+     * Retain file when delete record from DB.
+     *
+     * @var bool
+     */
+    protected $retainable = false;
+
+    /**
      * Configuration for setting up file actions for newly selected file thumbnails in the preview window.
      *
      * @var array
@@ -59,7 +66,7 @@ trait UploadField
      *
      * @var string
      */
-    protected $storage_permission;
+    protected $storagePermission;
 
     /**
      * Initialize the storage instance.
@@ -124,6 +131,18 @@ trait UploadField
     public function removable()
     {
         $this->fileActionSettings['showRemove'] = true;
+
+        return $this;
+    }
+
+    /**
+     * Indicates if the underlying field is retainable.
+     *
+     * @return $this
+     */
+    public function retainable($retainable = true)
+    {
+        $this->retainable = $retainable;
 
         return $this;
     }
@@ -297,8 +316,8 @@ trait UploadField
     {
         $this->renameIfExists($file);
 
-        if (!is_null($this->storage_permission)) {
-            return $this->storage->putFileAs($this->getDirectory(), $file, $this->name, $this->storage_permission);
+        if (!is_null($this->storagePermission)) {
+            return $this->storage->putFileAs($this->getDirectory(), $file, $this->name, $this->storagePermission);
         }
 
         return $this->storage->putFileAs($this->getDirectory(), $file, $this->name);
@@ -361,15 +380,15 @@ trait UploadField
     {
         $index = 1;
         $extension = $file->getClientOriginalExtension();
-        $originalName = $file->getClientOriginalName();
-        $newName = $originalName.'_'.$index.'.'.$extension;
+        $original = $file->getClientOriginalName();
+        $new = sprintf('%s_%s.%s', $original, $index, $extension);
 
-        while ($this->storage->exists("{$this->getDirectory()}/$newName")) {
+        while ($this->storage->exists("{$this->getDirectory()}/$new")) {
             $index++;
-            $newName = $originalName.'_'.$index.'.'.$extension;
+            $new = sprintf('%s_%s.%s', $original, $index, $extension);
         }
 
-        return $newName;
+        return $new;
     }
 
     /**
@@ -381,22 +400,27 @@ trait UploadField
     {
         if (is_array($this->original)) {
             foreach ($this->original as $item) {
-                if ($this->storage->exists($item)) {
+                if (!$this->retainable && $this->storage->exists($item)) {
                     $this->storage->delete($item);
                 }
             }
         } else {
-            if ($this->storage->exists($this->original)) {
+            if (!$this->retainable && $this->storage->exists($this->original)) {
                 $this->storage->delete($this->original);
             }
         }
-
-
     }
 
-    public function storage_permission($permission)
+    /**
+     * Set file permission when stored into storage.
+     *
+     * @param string $permission
+     *
+     * @return $this
+     */
+    public function storagePermission($permission)
     {
-        $this->storage_permission = $permission;
+        $this->storagePermission = $permission;
 
         return $this;
     }
